requirements.txt
streamlit
google-generativeai
st-gsheets-connection
pandas
extra-streamlit-components
python-docx
